feat(.c/.h//player/player_info//serialization//sprites): fin de la réorganisation des includes et du recadrage des scopes
