+++ base_map/base_map_manager.c
@@ -11,22 +11,20 @@
 
 #include <stdlib.h>
 #include <stdio.h>
-#include <SDL2/SDL.h>
-#include <SDL2/SDL_image.h>
-#include <SDL2/SDL_ttf.h>
 #include <pthread.h>
 #include "enum.h"
 #include "sdl.h"
 #include "player_info.h"
-#include "request.h"
+#include "client_request.h"
 #include "server.h"
 #include "player.h"
 #include "map.h"
 #include "data.h"
-#include "base_map_manager.h"
 #include "draw_base_map.h"
 #include "draw_players.h"
 #include "game_info.h"
+#include "game_map.h"
+#include "base_map_manager.h"
 
 // wrapper to init an SDL_Rect
 SDL_Rect init_rect(int x, int y, int w, int z)
