+++ client/client_request.c
@@ -10,7 +10,7 @@
 
 #include "sdl.h"
 #include "enum.h"
-#include "request.h"
+#include "client_request.h"
 #include "client.h"
 
 void	my_bzero(void *s1, int n)
@@ -59,22 +59,19 @@ int         send_request(int s, t_player_request* client_request)
   return SUCCESS_SEND;
 }
 
-int		get_msg(int s)
+// return a unique int corresponding to request values
+int     get_request_checksum(t_player_request* client_request)
 {
-  char		buff[1024];
-  int		r;
+  int   checksum = 0;
+  int   i;
+  unsigned char *p = (unsigned char *)&client_request->magic;
 
-  my_bzero(buff, 1024);
-  r = recv(s, buff, 1024, 0);
-  if (r > 0)
+  for (i = 0; i<(int)sizeof(client_request->magic); i++)
     {
-      buff[r] = '\0';
-      write(1, buff, r);
-      return SUCCESS_RECEIVE;
+      checksum += p[i];                 
     }
-  else
-    {
-      write(1, "Connection closed\n", 18);
-      return BAD_RECEIVE;
-    }
+  p = (unsigned char *)&client_request->command;
+  for (i = 0; i<(int)sizeof(client_request->command); i++)
+    checksum += p[i];  
+  return checksum;                   
 }
