+++ common/game_info_serialization.c
@@ -8,11 +8,13 @@
 ** Last update Wed Jul  4 09:33:55 2018 MASERA Mathieu
 */
 
+#include "enum.h"
 #include "sdl.h"
 #include "player_info.h"
-#include "request.h"
+#include "client_request.h"
 #include "server.h"
 #include "my_put.h"
+#include "map.h"
 #include "game_info.h"
 #include "game_info_serialization.h"
