+++ includes/base_map_manager.h
@@ -12,20 +12,6 @@
 #ifndef _BASE_MAP_MANAGER_H_
 #define _BASE_MAP_MANAGER_H_
 
-// need factorisation
-
-//game_map/white_bomber.c
-void *white_bomber_sprite(void *arg);
-//game_map/black_bomber.c
-void *black_bomber_sprite(void *arg);
-//game_map/red_bomber.c
-void *red_bomber_sprite(void *arg);
-//game_map/blue_bomber.c
-void *blue_bomber_sprite(void *arg);
-
-//game_map/bomb.c
-void *bomb_sprite(void *arg);
-
 SDL_Rect init_rect(int x, int y, int w, int z);
 t_map init_t_map(SDL_Rect src, SDL_Rect dest, texture_type type);
 void *draw_all(void *arg);
