+++ includes/client.h
@@ -9,11 +9,10 @@
 */
 
 #ifndef _CLIENT_H_
-#define _CLIEN_H_
+#define _CLIENT_H_
 
 void	init_client(t_sdl *sdl);
 char	*enter_addr(t_sdl *sdl);
-int	host_connect();
-int	client_connect();
-int	get_message(int s);
-#endif
+int	client_connect(char *serv_addr);
+
+#endif /* !_CLIENT_H_ */
