+++ includes/client_receive.h
@@ -1,0 +1,7 @@
+#ifndef _CLIENT_RECEIVE_H_
+#define _CLIENT_RECEIVE_H_
+
+void *listen_server(void *s);
+int  get_message(int s);
+
+#endif /* !_CLIENT_RECEIVE_H_ */
