+++ includes/game_map.h
@@ -1,0 +1,10 @@
+#ifndef _GAME_MAP_H_
+#define _GAME_MAP_H_
+
+void	*black_bomber_sprite(void *arg);
+void	*blue_bomber_sprite(void *arg);
+void	*white_bomber_sprite(void *arg);
+void	*red_bomber_sprite(void *arg);
+void	*bomb_sprite(void *arg);
+
+#endif /* !_GAME_MAP_H_ */
