+++ includes/moving.h
@@ -1,0 +1,7 @@
+#ifndef _MOVING_H_
+#define _MOVING_H_
+
+void	move_player(t_game_info *game_info, t_player_request *player_request, int num_player);
+// int	check_collision(int **map_pointer, t_player_request* player_request);
+
+#endif /* !_MOVING_H_ */
