+++ sdl.c
@@ -10,23 +10,24 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include "constant.h"
-#include "sdl.h"
 #include "enum.h"
+// doit être include au départ, défini SDL/* et t_sdl
+#include "sdl.h"
+//
 #include "map.h"
 #include "player_info.h"
-#include "request.h"
+#include "client_request.h"
 #include "server.h"
 #include "libmy.h"
 #include "base_map_manager.h"
 
 //On init la structure avec tous ses champs a NULL
-t_sdl 		*init_sdl()
+t_sdl	*init_sdl()
 {
-  t_sdl 		*sdl;
+  t_sdl	*sdl;
 
   if ((sdl = malloc(sizeof(t_sdl))) == NULL)
     return NULL;
-
   sdl->window = NULL;
   sdl->renderer = NULL;
   sdl->menu_background = NULL;
@@ -34,15 +35,14 @@ t_sdl 		*init_sdl()
   sdl->create_game = NULL;
   sdl->server_welcome = NULL;
   sdl->white_back = NULL;
-
   return sdl;
 }
 
 //Init de la window et du renderer
-t_sdl *init_window(t_sdl *sdl)
+t_sdl		*init_window(t_sdl *sdl)
 {
-  SDL_Window *window;
-  SDL_Renderer *renderer;
+  SDL_Window	*window;
+  SDL_Renderer	*renderer;
 
   window = SDL_CreateWindow(
 			    "Bomberman ETNA 2019",
@@ -51,24 +51,19 @@ t_sdl *init_window(t_sdl *sdl)
 			    WINDOW_W,
 			    WINDOW_H,
 			    SDL_WINDOW_SHOWN);
-
   if (!window)
     {
       printf("window init error");
       return NULL;
     }
-
   renderer = SDL_CreateRenderer(window, -1, 0);
-
   if (!renderer)
     {
       SDL_ShowSimpleMessageBox(0, "renderer init error", SDL_GetError(), window);
       return NULL;
     }
-
   sdl->window = window;
   sdl->renderer = renderer;
-
   return sdl;
 }
