+++ server/map_management.c
@@ -13,9 +13,12 @@
 #include "sdl.h"
 #include "enum.h"
 #include "player_info.h"
-#include "request.h"
+#include "client_request.h"
 #include "server.h"
+#include "map.h"
 #include "game_info.h"
+#include "bomb_management.h"
+#include "map_management.h"
 
 int		**get_array_map() {
   static int	map[104][88];
@@ -66,12 +69,10 @@ void			add_destructible_elements(
 }
 
 
-// détermine si l'on peut placer une bombe,
-// si une bombe est à l'emplacement, détermine si
-void			add_bomb_elements(
-					  t_game_info *game_info,
-					  int **map_pointer
-					  )
+void		        manage_bombs(
+				    t_game_info *game_info,
+				    int **map_pointer
+				    )
 {
   int			i;
   int			j;
@@ -125,7 +126,7 @@ int 			is_there_a_wall(
   int			i;
   int			j;
   t_map_destroyable	map_destroyable;
-
+  
   if (map_pointer[x][y] == WALL)
     return 1;
   for (i = 1; i < 14; ++i) {
@@ -138,6 +139,5 @@ int 			is_there_a_wall(
 	return 1;
     }
   }
-
   return 0;
 }
