+++ server/moving.c
@@ -13,9 +13,11 @@
 #include "enum.h"
 #include "sdl.h"
 #include "player_info.h"
-#include "request.h"
+#include "client_request.h"
 #include "server.h"
+#include "map.h"
 #include "game_info.h"
+#include "moving.h"
 
 void	move_player(
 		    t_game_info *game_info,
