+++ server/player.c
@@ -13,9 +13,8 @@
 #include <pthread.h>
 #include "enum.h"
 #include "sdl.h"
-#include "socket.h"
 #include "player_info.h"
-#include "request.h"
+#include "client_request.h"
 #include "server.h"
 #include "map.h"
 #include "data.h"
