+++ server/server.c
@@ -14,13 +14,11 @@
 #include "enum.h"
 #include "sdl.h"
 #include "player_info.h"
-#include "request.h"
+#include "client_request.h"
 #include "map.h"
 #include "data.h"
 #include "server.h"
-#include "socket.h"
 #include "thread.h"
-#include "request.h"
 #include "game_info.h"
 #include "my_put.h"
